requirements.txt
pymongo
requests
huggingface_hub
numpy
python-dotenv

